src/helpers/env.py
from decouple import Config, RepositoryEnv
from functools import lru_cache
from pathlib import Path

BASE_DIR = Path(__file__).resolve().parent.parent
BASE_DIR_ENV = BASE_DIR / '.env'


@lru_cache()
def get_config():
    if BASE_DIR_ENV.exists():
        return Config(RepositoryEnv(str(BASE_DIR_ENV)))

    from decouple import config
    return config


config = get_config()
